requirements.txt
fastapi==0.128.4
uvicorn==0.40.0
google-generativeai>=0.8.0
python-multipart==0.0.22
pillow>=12.0.0
python-dotenv==1.2.1
pydantic>=2.5.0
pydantic-settings>=2.1.0
gTTS>=2.5.1
gunicorn>=21.2.0
